feat:create migration address_shippings

+++ database/migrations/2026_07_21_065657_create-shipping_addresses-table.php
@@ -20,9 +20,8 @@ return new class extends Migration
             $table->integer('producer_id')->unsigned();
             $table->foreign("producer_id")->references("id")->on("producers")->onUpdate("cascade")->onDelete("cascade");
 
-            $table->integer('customer_id')->unsigned();
+            $table->bigInteger('customer_id')->unsigned();
             $table->foreign("customer_id")->references("id")->on("customers")->onUpdate("cascade")->onDelete("cascade");
-
             $table->timestamps();
         });
     }
